% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymise.R
\name{anonymise}
\alias{anonymise}
\title{Anonymise tidyxl data}
\usage{
anonymise(tbl, skip_rows = NULL, big_mark = ".", decimal_mark = ",")
}
\arguments{
\item{tbl}{tidyxl tibble}

\item{skip_rows}{integer vector of rows to skip}

\item{big_mark}{passed to \code{\link[=is_numeric_like]{is_numeric_like()}}}

\item{decimal_mark}{passed to \code{\link[=is_numeric_like]{is_numeric_like()}}}
}
\value{
tidyxl tibble
}
\description{
Make all characters anonymous, ignoring numbers, dates, and numbers
stored as characters.
}
\examples{
tidyxl::xlsx_cells("data_raw/pcontas.xlsx") |>
 anonymise(
   skip_rows = c(
     2:4, 12:14, 16, 89, 91, 229, 231, 233, 235:237, 252:254,
     264:266, 389,391:400, 402:403, 405:406, 408:410)) |>
 unpivotr::rectify()

}
