% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_partition_funs.R
\name{call_partition_funs}
\alias{call_partition_funs}
\title{Call Partition Functions}
\usage{
call_partition_funs(
  tbl,
  .code_col = code,
  .cells_col = cells,
  .progress = FALSE
)
}
\arguments{
\item{tbl}{partitioned tibble with required cols}

\item{.code_col}{code column to pass to \code{call_partition_funs()}}

\item{.cells_col}{cells column}

\item{.progress}{progress bar, default \code{TRUE}}
}
\value{
paritioned cells
}
\description{
Call Partition Functions
}
\examples{
parts <- system.file("data_raw/pcontas.xlsx", package = "tidycondr") |>
 read_contas() |>
 partition_contas()

parts[grepl("016", parts$code),] |>
 call_partition_funs(.code_col = code, .cells_col = cells)

}
