% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_row.R
\name{get_group_row}
\alias{get_group_row}
\title{Get Group Row by Position Id}
\usage{
get_group_row(cells, row, col = 1, col_info = "character")
}
\arguments{
\item{cells}{Tidy cells.}

\item{row}{Row number.}

\item{col}{Col number.}

\item{col_info}{Type of, as column name in tidy cells.}
}
\value{
Character vector.
}
\description{
Get Group Row by Position Id
}
