% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partition_corners_rows.R
\name{get_partition_corners_rows}
\alias{get_partition_corners_rows}
\title{Get Rows for Corners Partitions}
\usage{
get_partition_corners_rows(tbl, filled_cols, empty_cols)
}
\arguments{
\item{tbl}{Tidyxl tibble}

\item{filled_cols}{numeric vector for non empty columns, using original data}

\item{empty_cols}{numeric vector for empty columns, using original data}
}
\value{
numeric vector for rows
}
\description{
Get Rows for Corners Partitions
}
