% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partition_fun.R
\name{get_partition_fun}
\alias{get_partition_fun}
\title{Get Partition Function from String}
\usage{
get_partition_fun(str)
}
\arguments{
\item{str}{code for cells block}
}
\value{
function as \verb{partition_*} replacing *
}
\description{
Get Partition Function from String
}
\examples{
parts <- system.file("extdata/pcontas.xlsx", package = "tidycondr") |>
 read_contas() |>
 partition_contas()

get_partition_fun("016B")(list(parts[1, 4][[1]][[1]]))

}
