% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_rows.R
\name{head_rows}
\alias{head_rows}
\title{Rows Indices for Sequence Cell Values}
\usage{
head_rows(head_values, tbl_to_extract_rows, verbose = TRUE)
}
\arguments{
\item{head_values}{Vector or list of vectors}

\item{tbl_to_extract_rows}{Tibble to retrive column names from}

\item{verbose}{Verbose for \code{dplyr::inner_join()} messages}
}
\value{
Rows indices numeric vector
}
\description{
Rows Indices for Sequence Cell Values
}
