% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_rows.R
\name{head_rows_base}
\alias{head_rows_base}
\title{Rows Indices for Sequence Cell Values}
\usage{
head_rows_base(head_values, tbl_to_extract_rows)
}
\arguments{
\item{head_values}{Vector or list of vectors}

\item{tbl_to_extract_rows}{Tibble to retrive column names from}
}
\value{
Rows indices numeric vector
}
\description{
Rows Indices for Sequence Cell Values
}
