% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_016.R
\name{partition_016}
\alias{partition_016}
\title{Process 016* Report Partition}
\usage{
partition_016(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Original tibble with new column \code{processed_data}
}
\description{
Process financial summary (Resumo financeiro).
}
\examples{
pcontas <- system.file("extdata/pcontas.xlsx", package = "tidycondr") |>
 read_contas()
pcontas_part <- partition_contas(pcontas)
pcontas_part <- pcontas_part[grepl("016", pcontas_part$code),]

partition_016(pcontas_part[1,4][[1]][[1]])

pcontas_part |>
 mutate(processed = purrr::map(cells, partition_016))

}
