% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_016b.R
\name{partition_016b}
\alias{partition_016b}
\title{Process 016B Report Partition}
\usage{
partition_016b(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Original tibble with new column \code{processed_data}
}
\description{
Process financial summary (Resumo financeiro).
}
\examples{
pcontas <- read_contas("ex/pcontas.xlsx")
pcontas_part <- partition_contas(pcontas)
pcontas_part <- filter(pcontas_part, code == "016B")

partition_016b(pcontas_part[1,4][[1]][[1]])

pcontas_part |>
 mutate(processed = purrr::map(cells, partition_016b))

}
