% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_020.R
\name{partition_020}
\alias{partition_020}
\title{Process 020 Report Partition}
\usage{
partition_020(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Process partitions of financial report:
\itemize{
\item Resumo movimentação
\item Demonstrativo de receitas e despesas analítico
\item Demonstrativo analítico c fundo de caixa
}
}
\examples{
pcontas <- read_contas("data_raw/pcontas.xlsx")
pcontas_part <- partition_contas(pcontas)
pcontas_part <- pcontas_part[grepl("020", pcontas_part$code),]

partition_020(pcontas_part[1,4][[1]][[1]])

pcontas_part |>
 mutate(processed = purrr::map(cells, partition_020))

}
