% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_020_despesas.R
\name{partition_020_despesas}
\alias{partition_020_despesas}
\title{Process 020 Despesas Analítico Report Partition}
\usage{
partition_020_despesas(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Process despesas partition of financial report
(Demonstrativo de receitas e despesas analítico |>
despesas).
}
