% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_020_groups.R
\name{partition_020_groups}
\alias{partition_020_groups}
\title{Partition 020 Report into Groups Receitas / Despesas}
\usage{
partition_020_groups(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Split groups Receitas / Despesas partition of financial report
(Demonstrativo de receitas e despesas analítico).
}
\examples{
pcontas <- system.file("extdata/pcontas.xlsx", package = "tidycondr") |>
 read_contas()
pcontas_part <- partition_contas(pcontas)
pcontas_part <- pcontas_part[grepl("020", pcontas_part$code),]

partition_020_groups(pcontas_part$cells[[1]])

}
