% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_020_summary.R
\name{partition_020_summary}
\alias{partition_020_summary}
\title{Process 020 Summary Report Partition}
\usage{
partition_020_summary(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Process summary partition of financial report
(Demonstrativo de receitas e despesas analítico).
}
\examples{
pcontas <- system.file("extdata/pcontas.xlsx", package = "tidycondr") |>
 read_contas()
pcontas_part <- partition_contas(pcontas)
pcontas_part <- pcontas_part[grepl("020", pcontas_part$code),]

partition_020_summary(pcontas_part[1,4][[1]][[1]])

pcontas_part |>
 dplyr::mutate(processed = purrr::map(cells, partition_020_summary))

}
