% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_028.R
\name{partition_028}
\alias{partition_028}
\title{Process 028 Report Partition}
\usage{
partition_028(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Process partitions of financial report:
\itemize{
\item Acordos
\item Cobranças originais do acordo
\item Parcelas do acordo
}
}
\examples{
acordos <- system.file("extdata/acordos.xlsx", package = "tidycondr") |>
  read_contas()
acordos_part <- partition_contas(acordos)
acordos_part <- acordos_part[grepl("028", acordos_part$code),]
partition_028(acordos_part[1,4][[1]][[1]])
acordos_part |>
  mutate(processed = purrr::map(cells, partition_028))

}
