% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_028_acordo_detail.R
\name{partition_028_acordo_detail}
\alias{partition_028_acordo_detail}
\title{Process 028* Report Partition Details}
\usage{
partition_028_acordo_detail(tbl_acordo)
}
\arguments{
\item{tbl_acordo}{acordo partition tibble header}
}
\value{
pivoted wider tibble for futuro unnest
}
\description{
Process 028* Report Partition Details
}
