% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_028_parcelas.R
\name{partition_028_parcelas}
\alias{partition_028_parcelas}
\title{Process 028 parcelas Report Partition}
\usage{
partition_028_parcelas(parc)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Process parcelas_do_acordo partition of financial report.
}
\examples{
acord <- read_contas("data_raw/acordos.xlsx")
pacord <- partition_contas(acord)
pacord <- pacord[grepl("028", pacord$code),]
pacord_part <- pacord$cells[1][[1]]
tbl <- partition_028_split_groups(pacord_part)
parc <- tbl$parcelas_do_acordo[1][[1]]
partition_028_parcelas(parc)

}
