% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_028_split_groups.R
\name{partition_028_split_groups}
\alias{partition_028_split_groups}
\title{Partition 028 Report into Groups Cobrancas Renegociadas / Parcelas Definidas}
\usage{
partition_028_split_groups(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy nested data.
}
\description{
Split groups Cobrancas Renegociadas / Parcelas Definidas partition of
financial report (cobrancas_originais e parcelas_do_acordo ).
}
\examples{
acord <- read_contas("data_raw/acordos.xlsx")
pacord_part <- partition_contas(acord)
pacord_part <- pacord_part[grepl("028", pacord_part$code),]

partition_028_split_groups(pacord_part[1,4][[1]][[1]])

pcontas_part |>
 mutate(processed = purrr::map(cells, partition_028_split_groups))

}
