% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_028_summary.R
\name{partition_028_summary}
\alias{partition_028_summary}
\title{Process 028* Report Partition Details}
\usage{
partition_028_summary(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy nested data.
}
\description{
Process financial summary (Acordo detalhado).
}
\examples{
acord <- read_contas("extdata/acordos.xlsx")
pacord_part <- partition_contas(acord)
pacord_part <- pacord_part[grepl("028", pacord_part$code),]

partition_028_summary(pacord_part[1,4][[1]][[1]])

pcontas_part |>
 mutate(processed = purrr::map(cells, partition_028_summary))

}
