% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_033.R
\name{partition_033}
\alias{partition_033}
\title{Process 033* Report Partition}
\usage{
partition_033(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Original tibble with new column \code{processed_data}
}
\description{
Process financial summary (Cheques emitidos).
}
