% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_037.R
\name{partition_037}
\alias{partition_037}
\title{Process 037* Report Partition}
\usage{
partition_037(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Original tibble with new column \code{processed_data}
}
\description{
Process financial summary (Resumo financeiro).
}
\examples{
pcontas <- read_contas("ex/pcontas.xlsx")
pcontas_part <- partition_contas(pcontas)
pcontas_part <- pcontas_part[grepl("037", parts$code),]

partition_037(pcontas_part[1,4][[1]][[1]])

pcontas_part |>
 mutate(processed = purrr::map(cells, partition_037))

}
