% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_039.R
\name{partition_039}
\alias{partition_039}
\title{Process 039* Report Partition}
\usage{
partition_039(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Process financial summary (Evolução de despesa anual).
}
\examples{
pcontas <- read_contas("data_raw/pcontas.xlsx")
pcontas_part <- partition_contas(pcontas)
pcontas_part <- pcontas_part[grepl("039", pcontas_part$code),]

partition_039(pcontas_part[1,4][[1]][[1]])

pcontas_part |>
 mutate(processed = purrr::map(cells, partition_039))

}
