% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_060.R
\name{partition_060}
\alias{partition_060}
\title{Process 060* Report Partition}
\usage{
partition_060(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy nested data.
}
\description{
Process financial summary (Demonstrativo de inadimplência, por unidade e anual).
}
\examples{
pcontas <- system.file("data_raw/pcontas.xlsx", package = "tidycondr") |>
 read_contas()
pcontas_part <- partition_contas(pcontas)
pcontas_part <- pcontas_part[grepl("060", pcontas_part$code),]

partition_060(pcontas_part[1,4][[1]][[1]])
partition_060(pcontas_part[2,4][[1]][[1]])

pcontas_part |>
 mutate(processed = purrr::map(cells, partition_060))

}
