% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_contas.R
\name{partition_contas}
\alias{partition_contas}
\title{Split Contas in Partitions}
\usage{
partition_contas(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=read_contas]{read_contas()}}.}
}
\value{
A tibble with nested cols, 1 row per report
}
\description{
Get \code{\link[unpivotr:partition]{unpivotr::partition()}} for each report in \code{\link[=read_contas]{read_contas()}} tibble.
}
\examples{
pcontas <- system.file("extdata/pcontas.xlsx", package = "tidycondr") |>
 read_contas()
partition_contas(pcontas)

}
