% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_title.R
\name{pull_info}
\alias{pull_info}
\alias{pull_title}
\alias{pull_subtitle}
\alias{pull_code}
\title{Pull Info from Partition}
\usage{
pull_title(df)

pull_subtitle(df)

pull_code(df)
}
\arguments{
\item{df}{A tidyxl tibble}
}
\value{
Title, subtitle or code character vector.
}
\description{
Pull info from a tidyxl from a tibble, previously partitioned.
}
