% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_contas.R
\name{read_contas}
\alias{read_contas}
\title{Contas xlsx Contents}
\usage{
read_contas(contas_file)
}
\arguments{
\item{contas_file}{xlsx filepath}
}
\value{
\code{\link[tidyxl:xlsx_cells]{tidyxl::xlsx_cells()}} tibble with selected variables.
}
\description{
Read \link{tidyxl} data from \code{xlsx}file.
}
\details{
Outputs tibble with selected variables using \code{\link[tidyxl:xlsx_cells]{tidyxl::xlsx_cells()}}:
\itemize{
\item \code{row} The row number of a cell address (integer).
\item \code{col} The column number of a cell address (integer).
\item \code{is_blank} Whether or not the cell has a value
\item \code{content} Raw cell value before type conversion, useful for debugging.
\item \code{data_type} The type of a cell, referring to the following columns: error,
logical, numeric, date, character, blank.
\item \code{error} The error value of a cell.
\item \code{logical} The boolean value of a cell.
\item \code{numeric} The numeric value of a cell.
\item \code{date} The date value of a cell.
\item \code{character} The string value of a cell.
}
}
\examples{
read_contas("data_raw/pcontas.xlsx")
read_contas("data_raw/acordos.xlsx")
}
\seealso{
\code{\link[tidyxl:xlsx_cells]{tidyxl::xlsx_cells()}} which this function wraps.
}
