% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_contas_partitions.R
\name{read_contas_partitions}
\alias{read_contas_partitions}
\title{Contas xlsx Partitions Contents}
\usage{
read_contas_partitions(contas_file, .progress = TRUE)
}
\arguments{
\item{contas_file}{xlsx filepath}

\item{.progress}{show {purrr} progress bar?}
}
\value{
tidy partitions.
}
\description{
Read \link{tidyxl} data from \code{xlsx}file and retrieve all partitions using
partition_*() for each report code.
}
\examples{
read_contas_partitions("data_raw/pcontas.xlsx", .progress = TRUE)
}
