% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dataset.R
\name{reduce_acordos}
\alias{reduce_acordos}
\title{Reduce acordos datasets tibble to one row per report}
\usage{
reduce_acordos(dataset)
}
\arguments{
\item{dataset}{tibble dataset from \code{\link[=read_contas_partitions]{read_contas_partitions()}}.}
}
\value{
tibble
}
\description{
Bind rows from tibbles in dataset. It reduces from multiple rows in dataset,
one row for each file
}
\examples{
read_acordos_partitions("extdata/acordos.xlsx", .progress = TRUE) |>
  reduce_acordos()

}
