% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_rand_case.R
\name{replace_rand_case}
\alias{replace_rand_case}
\title{Replace Characters with Random Characters}
\usage{
replace_rand_case(string)
}
\arguments{
\item{string}{Atomic string vector}
}
\value{
Atomic string vector
}
\description{
Replace chars in string with random chars with case matching.
}
\examples{
str <- "String a ser testada Com NOME 12456"
replace_rand_case(str)

str_list <- str_list <- list(
 "String a ser testada Com NOME 12456",
 "String COM data: 13/06/1985",
 "Outra STRING com número e $pecial CHAR #$\%")
 replace_rand_case(str_list[[1]])
 lapply(str_list, replace_rand_case)
 sapply(str_list, replace_rand_case)

}
