% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycondr-package.R
\docType{package}
\name{tidycondr-package}
\alias{tidycondr}
\alias{tidycondr-package}
\title{tidycondr: Tidy Condo Data}
\description{
Tidy messy data from Superlogica `xlsx` outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/diegomsg/tidycondr}
  \item Report bugs at \url{https://github.com/diegomsg/tidycondr/issues}
}

}
\author{
\strong{Maintainer}: Diego Guimarães \email{diegomsg@gmail.com}

}
\keyword{internal}
