% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_rows.R
\name{vec_to_tibl}
\alias{vec_to_tibl}
\title{Retrive Column Names from Tibble}
\usage{
vec_to_tibl(vec, names_from_tbl)
}
\arguments{
\item{vec}{Character vector sequence of cell values in row}

\item{names_from_tbl}{Tibble to retrive column names from}
}
\value{
Tibble with col names and values in vector.
}
\description{
Retrive Column Names from Tibble
}
