% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tidyxl.R
\name{check_tidyxl}
\alias{check_tidyxl}
\title{Check a tidyxl tibble}
\usage{
check_tidyxl(tbl)
}
\arguments{
\item{tbl}{tibble}
}
\value{
logical
}
\description{
Check if it`s a simple \code{\link[tidyxl:xlsx_cells]{tidyxl::xlsx_cells()}} with selected cols.
}
\examples{
pcontas <- read_contas("data_raw/pcontas.xlsx")
check_tidyxl(pcontas)

acordos <- read_contas("data_raw/acordos.xlsx")
check_tidyxl(acordos)
}
