% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_020_despesas.R
\name{partition_020_despesas}
\alias{partition_020_despesas}
\title{Process 020 Despesas Analítico Report Partition}
\usage{
partition_020_despesas(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Process despesas partition of financial report
(Demonstrativo de receitas e despesas analítico |>
despesas).
}
\examples{
pcontas <- system.file("data_raw/pcontas.xlsx", package = "tidycondr") |>
 read_contas()
pcontas_part <- partition_contas(pcontas)
pcontas_part <- pcontas_part[grepl("020", pcontas_part$code),]

tbl_groups <- partition_020_groups(tbl)

partition_020_despesas(
 tbl_groups[tbl_groups$info == "Despesas",]$data[[1]])

}
