% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_020_receitas.R
\name{partition_020_receitas}
\alias{partition_020_receitas}
\title{Process 020 Receitas Analítico Report Partition}
\usage{
partition_020_receitas(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Process receitas partition of financial report
(Demonstrativo de receitas e despesas analítico |>
receitas).
}
