% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_028.R
\name{partition_028}
\alias{partition_028}
\title{Process 028 Report Partition}
\usage{
partition_028(tbl)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Process partitions of financial report:
\itemize{
\item Acordos
\item Cobranças originais do acordo
\item Parcelas do acordo
}
}
\examples{
pcontas <- system.file("extdata/pcontas.xlsx", package = "tidycondr") |>
 read_contas()
pcontas_part <- partition_contas(pcontas)
pcontas_part <- pcontas_part[grepl("028", pcontas_part$code),]

partition_028(pcontas_part[1,4][[1]][[1]])

pcontas_part |>
 mutate(processed = purrr::map(cells, partition_028))

}
