% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_028_cobrancas.R
\name{partition_028_cobrancas}
\alias{partition_028_cobrancas}
\title{Process 028 Cobrancas Report Partition}
\usage{
partition_028_cobrancas(cob)
}
\arguments{
\item{tbl}{A tidyxl tibble from \code{\link[=partition_contas]{partition_contas()}}}
}
\value{
Tidy data.
}
\description{
Process cobrancas_originais partition of financial report.
}
\examples{
acordos <- system.file("extdata/acordos.xlsx", package = "tidycondr") |>
  read_contas()
pacord <- partition_contas(acordos)
pacord <- pacord[grepl("028", pacord$code),]
pacord_part <- pacord$cells[1][[1]]
tbl <- partition_028_split_groups(pacord_part)
cob <- tbl$cobrancas_originais[1][[1]]
partition_028_cobrancas(cob)

}
